PG-228: Adjusted error messages.
Insufficient shared space and error when reading queries from disk must
be handled separately.

If we try to read query from memory and fail to find it, then it is
because there was not enough memory in the space allocated for query
buffer using the pgsm_query_shared_buffer GUC, so, at some point this
buffer was cleared for storing most recent queries in favor of old ones.

But if pgsm_overflow_target GUC is set to 1 (dump queries to disk when
the query buffer would overflow), then we attempt to read the query
from disk, if that fails it means some I/O error happened.