PG-2116 Truncate query strings at multi-byte character boundaries
Make sure we never create broken encodings in query strings by making
the truncation of query strings multi-byte aware. An alternative would
be to change the rule for truncation to be based on number of
characters, as the documentation claims it is, but I do not think that
would be significantly better. Either works just fine so we just pick
one.

Based on bug report from Ivan Vyazmitinov:

https://github.com/percona/pg_stat_monitor/issues/599